Airflow/requirements.txt
apache-airflow-providers-apache-kafka>=1.7.0
confluent-kafka